// interfaces/SidebarInterface.ts
// Interface para os itens de navegação
export interface NavItemInterface {
  label: string;
  href: string;
  icon: React.ReactNode;
}

// Interface principal do Sidebar
export interface SidebarInterface {
  // Por enquanto o Sidebar não recebe props, mas podemos expandir conforme necessário
} 